Version 0.1.1.5.6 - Alpha
Changelog:

Added:

+ Monsters can now jump higher, take no fall damage, spawn with weapons
and have a model.
+ Humans now have a model.
+ Screen message when a player turns into a monster.

Improved:

+ monsterManager runs on spawn.
+ Cvars actually do what they're supposed to now.
+ mt_forcemonster has crude admin protection.

--------------------------------------------------------------------------------
Dev Changelog:

Added:

+ Monsters can now jump higher, take no fall damage, spawn with weapons
and have a model.
+ Humans now have a model.
+ Screen message when a player turns into a monster.
+ Entity emitted sounds and screen shaking can be called from serverside
(not sure if this will last).

Improved:

+ Cvars actually do what they're supposed to now.
+ mt_forcemonster has crude admin protection.

Deleted:

+ Clientside monsterManager has been disabled.

Other:

+ pickMonster renamed to forceMonster.
+ setupMonster has been merged with unsetupMonster.
+ Moved clientside player functions to cl_player.
+ monsterMaker has been renamed to monsterManager.
+ monsterManager runs on spawn.

Notes:

This is a very messy build, I want to change a lot of things. Despite that it should be bug-free. Two bugs have been spotted in this build, but they aren't the worst.

In cl_init the code within GM:Think can calculate if a player is moving on the ground or not. Might be useful at some point.